refactor: split files over the 800-line limit
Split DataGrid.tsx, BottomPanel.tsx, Sidebar.tsx, and state.rs by
responsibility; no behavior or API changes. Document the 800-line rule
in CLAUDE.md and clarify the generated-file exemption in AGENTS.md.
