+++ CLAUDE.md
@@ -1,0 +1,9 @@
+# CLAUDE.md
+
+Follow `AGENTS.md` — it is the canonical guidance for agents in this repo.
+
+## File Size Limit
+
+- No human-authored source, documentation, or configuration file may exceed **800 lines**.
+- If a file approaches that size, split it by responsibility before adding more code.
+- Generated files (e.g. `packages/ipc/src/generated.ts`, lockfiles) and binary assets are exempt, but never hand-edit them — regenerate through their owning tools.
